requirements.txt
openai==1.6.1
langchain==0.1.9
chromadb==0.4.22
pymupdf==1.23.5
python-dotenv==1.0.1
streamlit==1.31.1
tiktoken==0.5.2
httpx==0.27
python-docx==1.1.2
python-pptx==1.0.2
langchain-openai==0.3.9